# 🚀 Improve performance for the Monix integration