--- a/core/src/main/scala/agni/std/Future.scala
+++ b/core/src/main/scala/agni/std/Future.scala
@@ -13,7 +13,7 @@ import com.google.common.util.concurrent.{ FutureCallback, Futures }
 import scala.concurrent.{ ExecutionContext, Promise, Future => SFuture }
 
 abstract class Future(implicit ec: ExecutionContext, _cache: Cache[String, PreparedStatement])
-    extends Agni[SFuture, Throwable] with CachedPreparedStatementWithGuava {
+  extends Agni[SFuture, Throwable] with CachedPreparedStatementWithGuava {
 
   override val F: MonadError[SFuture, Throwable] = catsStdInstancesForFuture