… core/src/main/scala/agni/RowDecoder.scala