This PR slightly improvements performance for the `getAsync` method of the Monix integrations.

<details>
<summary>Benchmarks (before and after)</summary>

**Before**

```
[info] Benchmark                                                   Mode  Cnt        Score        Error   Units
[info] MonixTaskBenchmark.one                                     thrpt   10     1238.810 ±    115.783   ops/s
[info] MonixTaskBenchmark.one:·gc.alloc.rate                      thrpt   10       20.309 ±      5.365  MB/sec
[info] MonixTaskBenchmark.one:·gc.alloc.rate.norm                 thrpt   10    20826.804 ±   3263.842    B/op
[info] MonixTaskBenchmark.one:·gc.churn.PS_Eden_Space             thrpt   10       23.033 ±     19.764  MB/sec
[info] MonixTaskBenchmark.one:·gc.churn.PS_Eden_Space.norm        thrpt   10    24031.376 ±  20348.512    B/op
[info] MonixTaskBenchmark.one:·gc.churn.PS_Survivor_Space         thrpt   10        0.011 ±      0.038  MB/sec
[info] MonixTaskBenchmark.one:·gc.churn.PS_Survivor_Space.norm    thrpt   10       10.991 ±     37.195    B/op
[info] MonixTaskBenchmark.one:·gc.count                           thrpt   10        8.000               counts
[info] MonixTaskBenchmark.one:·gc.time                            thrpt   10       46.000                   ms
[info] MonixTaskBenchmark.three                                   thrpt   10      401.915 ±     21.572   ops/s
[info] MonixTaskBenchmark.three:·gc.alloc.rate                    thrpt   10      374.290 ±    105.591  MB/sec
[info] MonixTaskBenchmark.three:·gc.alloc.rate.norm               thrpt   10  1174860.667 ± 152574.780    B/op
[info] MonixTaskBenchmark.three:·gc.churn.PS_Eden_Space           thrpt   10      390.425 ±    112.535  MB/sec
[info] MonixTaskBenchmark.three:·gc.churn.PS_Eden_Space.norm      thrpt   10  1232678.045 ± 203309.184    B/op
[info] MonixTaskBenchmark.three:·gc.churn.PS_Survivor_Space       thrpt   10        0.157 ±      0.096  MB/sec
[info] MonixTaskBenchmark.three:·gc.churn.PS_Survivor_Space.norm  thrpt   10      525.915 ±    477.554    B/op
[info] MonixTaskBenchmark.three:·gc.count                         thrpt   10       56.000               counts
[info] MonixTaskBenchmark.three:·gc.time                          thrpt   10       81.000                   ms
```

**After**

```
[info] Benchmark                                                   Mode  Cnt        Score        Error   Units
[info] MonixTaskBenchmark.one                                     thrpt   10     1550.227 ±     63.135   ops/s
[info] MonixTaskBenchmark.one:·gc.alloc.rate                      thrpt   10       25.706 ±      7.809  MB/sec
[info] MonixTaskBenchmark.one:·gc.alloc.rate.norm                 thrpt   10    20846.435 ±   3251.766    B/op
[info] MonixTaskBenchmark.one:·gc.churn.PS_Eden_Space             thrpt   10       25.804 ±     21.158  MB/sec
[info] MonixTaskBenchmark.one:·gc.churn.PS_Eden_Space.norm        thrpt   10    20212.258 ±  16477.675    B/op
[info] MonixTaskBenchmark.one:·gc.churn.PS_Survivor_Space         thrpt   10        0.012 ±      0.031  MB/sec
[info] MonixTaskBenchmark.one:·gc.churn.PS_Survivor_Space.norm    thrpt   10        9.762 ±     23.946    B/op
[info] MonixTaskBenchmark.one:·gc.count                           thrpt   10        8.000               counts
[info] MonixTaskBenchmark.one:·gc.time                            thrpt   10       39.000                   ms
[info] MonixTaskBenchmark.three                                   thrpt   10      421.349 ±     26.664   ops/s
[info] MonixTaskBenchmark.three:·gc.alloc.rate                    thrpt   10      390.921 ±    109.727  MB/sec
[info] MonixTaskBenchmark.three:·gc.alloc.rate.norm               thrpt   10  1172252.940 ± 162208.430    B/op
[info] MonixTaskBenchmark.three:·gc.churn.PS_Eden_Space           thrpt   10      400.750 ±     99.587  MB/sec
[info] MonixTaskBenchmark.three:·gc.churn.PS_Eden_Space.norm      thrpt   10  1216783.951 ± 229042.414    B/op
[info] MonixTaskBenchmark.three:·gc.churn.PS_Survivor_Space       thrpt   10        0.138 ±      0.109  MB/sec
[info] MonixTaskBenchmark.three:·gc.churn.PS_Survivor_Space.norm  thrpt   10      426.011 ±    336.078    B/op
[info] MonixTaskBenchmark.three:·gc.count                         thrpt   10       57.000               counts
[info] MonixTaskBenchmark.three:·gc.time                          thrpt   10       84.000                   ms
```
</details>